wip(fix): restrict wire1 inputs to unitary, poorly
In order to achieve the goal of 2c94850 without too much wrangling, we
restrict things that are inputs into a node to only produce a single
future scheduled event so we can avoid needing to possibly call
downstream consumers more than once in the future.

Alternatives include:
* Possibly restricting consumers to being Fn (vs FnMut), and passing
  around lots of references with *hand wave* the correct lifetimes
* Instead of taking consumers, take a FnMut that creates the consumers,
  (which can then all be FnOnces again), calling as many times as
  necessary. Also, it'd probably have to return a copy of itself for
  "rewiring", too

But probably the most fruitful:
* Stop holding Rust like it was Haskell, and make `struct Node {...}`
  that holds the (mutable) state of a node in the graph. (why is it
  always graphs?)

Additionally, and more reason to do the above, the simulator engine now
has to recognize the mapping between the schedule (in its entirety) and
which nodes are updated at which point for saving into the history. This
is probably fine (if not ideal) for a ~5 component graph, but unworkable
for a general simulator.